harvester/requirements.txt
lxml==3.4.1
inflection==0.3.1
pandas==0.19.2
elasticsearch==5.2.0
kafka-python==1.3.3
